refactor: move group execution logic to ctxrunner.py
